+++ src/consts.rs
@@ -3,18 +3,19 @@ pub use libsecp256k1::util::COMPRESSED_PUBLIC_KEY_SIZE;
 /// Uncompressed public key size
 pub use libsecp256k1::util::FULL_PUBLIC_KEY_SIZE as UNCOMPRESSED_PUBLIC_KEY_SIZE;
 
-/// AES nonce length
-#[cfg(not(feature = "aes-12bytes-nonce"))]
-pub const AES_NONCE_LENGTH: usize = 16;
-#[cfg(feature = "aes-12bytes-nonce")]
-pub const AES_NONCE_LENGTH: usize = 12;
-
-/// XChaCha20 nonce length
+/// Nonce length. AES (12/16 bytes) or XChaCha20 (24 bytes)
+#[cfg(all(not(feature = "aes-12bytes-nonce"), not(feature = "xchacha20")))]
+pub const NONCE_LENGTH: usize = 16;
+#[cfg(all(feature = "aes-12bytes-nonce", not(feature = "xchacha20")))]
+pub const NONCE_LENGTH: usize = 12;
 #[cfg(feature = "xchacha20")]
-pub const XCHACHA20_NONCE_LENGTH: usize = 24;
+pub const NONCE_LENGTH: usize = 24;
 
 /// AEAD tag length
 pub const AEAD_TAG_LENGTH: usize = 16;
+
+/// Nonce + tag length
+pub const NONCE_TAG_LENGTH: usize = NONCE_LENGTH + AEAD_TAG_LENGTH;
 
 /// Empty bytes array
 pub const EMPTY_BYTES: [u8; 0] = [];